# src/main/resources/static/openapi.yaml
openapi: 3.2.0

info:
  title: miscale2pg API
  version: 0.2.0
  summary: Import Mi Body Composition Scale 2 measurements into PostgreSQL
  license:
    name: MIT
    url: https://github.com/nasenov/miscale2pg/blob/main/LICENSE

tags:
  - name: measurements
  - name: actuator

paths:
  /api/measurements:
    get:
      operationId: getMeasurements
      summary: Get measurements by time range
      description: >-
        Get measurements within the given time range.
        If time range is omitted it defaults to last 30 days.
      tags:
        - measurements
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
          example: "2026-06-29T00:00:00Z"
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
          example: "2026-07-29T00:00:00Z"
      responses:
        "200":
          description: Measurements within the given time range
          content:
            application/json:
              example:
                - time: "2026-07-28T06:46:22Z"
                  weight: 66.6
                  height: 180.0
                  bmi: 20.5
                  fatRate: 14.14
                  bodyWaterRate: 58.9
                  boneMass: 2.91
                  metabolism: 1489.0
                  muscleRate: 54.27
                  visceralFat: 6.0

    post:
      operationId: uploadMeasurements
      summary: Upload measurements
      tags:
        - measurements
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  description: >-
                    Password-protected ZIP file exported from https://user.huami.com/privacy/index.html 
                    or CSV file(s) extracted from the password-protected ZIP file
                  type: string
                  format: binary
                password:
                  description: Password for the password-protected ZIP file
                  type: string
      responses:
        "201":
          description: Successful upload

  /api/measurements/sync:
    post:
      operationId: syncMeasurements
      summary: Synchronize measurements from Mi Fit / Zepp Life
      tags:
        - measurements
      parameters:
        - name: App-Token
          in: header
          description: >-
            Mi Fit / Zepp Life API authentication token
          required: true
          schema:
            type: string
        - name: userId
          in: query
          description: Mi Fit / Zepp Life user id
          required: true
          schema:
            type: string
        - name: email
          in: query
          description: >-
            Email address associated with the Mi Fit / Zepp Life account
          required: true
          schema:
            type: string
            format: email
        - name: startDate
          in: query
          description: >-
            If omitted, the date of the latest measurement already stored in the
            database is used. If the database is empty, endDate - 30 days date is used
          required: false
          schema:
            type: string
            format: date
          example: "2026-08-10"
        - name: endDate
          in: query
          description: >-
            If omitted, today date is used
          required: false
          schema:
            type: string
            format: date
          example: "2026-08-17"
      responses:
        "201":
          description: Successful sync

  /actuator/health:
    get:
      operationId: health
      summary: Application health
      tags:
        - actuator
      responses:
        "200":
          description: Application is healthy
        "503":
          description: Application is unhealthy

  /actuator/prometheus:
    get:
      operationId: prometheus
      summary: Prometheus metrics
      tags:
        - actuator
      responses:
        "200":
          description: Application metrics in Prometheus format